Separate service request from service lookup
rename WorkflowTrackingServiceEndpoints to Endpoints
extract class for service lookup
