+++ src/main/groovy/life/qbic/portal/infrastructure/services/workflowtracking/Endpoints.groovy
@@ -1,6 +1,6 @@
 package life.qbic.portal.infrastructure.services.workflowtracking
 
-class WorkflowTrackingServiceEndpoints {
+class Endpoints {
 
     final public static def WORKFLOW_LIST = "/workflows"
     final public static def WORKFLOW_INFO = "/workflows/info/"
